i2pProbe/pyproject.toml
[project]
name = "i2pProbe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "fastapi>=0.115.12",
    "observlib",
    "opentelemetry-instrumentation-aiohttp-client>=0.55b1",
    "opentelemetry-instrumentation-asyncio>=0.55b1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "prometheus-client>=0.22.0",
    "uvicorn>=0.34.2",
]

[project.scripts]
i2pProbe = "i2pProbe:__main__.run"

[tool.uv.sources]
observlib = { git = "https://github.com/ForgottenBeast/observlib.git", subdirectory = "observlib" }
